과제 2번 완료
